+++ src/main/kotlin/com/rtomyj/podcast/exception/PodcastError.kt
@@ -1,3 +1,3 @@
-package com.rtomyj.podcast.api.exception
+package com.rtomyj.podcast.exception
 
 data class PodcastError(val message: String, val code: String)
