+++ src/main/kotlin/com/rtomyj/podcast/exception/PodcastException.kt
@@ -1,6 +1,6 @@
-package com.rtomyj.podcast.api.exception
+package com.rtomyj.podcast.exception
 
-import com.rtomyj.podcast.api.util.enum.ErrorType
+import com.rtomyj.podcast.util.enum.ErrorType
 
 data class PodcastException(override val message: String, val errorType: ErrorType)
 	: RuntimeException()
