+++ src/main/kotlin/com/rtomyj/podcast/model/PodcastData.kt
@@ -1,3 +1,3 @@
-package com.rtomyj.podcast.api.model
+package com.rtomyj.podcast.model
 
 data class PodcastData(val podcast: Podcast, val podcastEpisodes: MutableList<PodcastEpisode>)
